Split deploy local into its own command.
